# api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/go-gost/x/config
    FileRecorder:
        properties:
            path:
                type: string
                x-go-name: Path
            sep:
                type: string
                x-go-name: Sep
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPRecorder:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
...
    LimiterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            limits:
                items:
                    type: string
                type: array
                x-go-name: Limits
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ObserverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    PluginConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
            tls:
                $ref: '#/definitions/TLSConfig'
            token:
                type: string
                x-go-name: Token
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    RecorderConfig:
        properties:
            file:
                $ref: '#/definitions/FileRecorder'
            http:
                $ref: '#/definitions/HTTPRecorder'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisRecorder'
            tcp:
                $ref: '#/definitions/TCPRecorder'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    RedisRecorder:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: github.com/go-gost/x/api
...
    TCPRecorder:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            options:
                $ref: '#/definitions/TLSOptions'
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    TLSOptions:
        properties:
            cipherSuites:
                items:
                    type: string
                type: array
                x-go-name: CipherSuites
            maxVersion:
                type: string
                x-go-name: MaxVersion
            minVersion:
                type: string
                x-go-name: MinVersion
        type: object
        x-go-package: github.com/go-gost/x/config
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
...
    /config/observers:
        post:
            operationId: createObserverRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ObserverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createObserverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new observer, the name of the observer must be unique in observer list.
            tags:
                - Observer
    /config/observers/{observer}:
        delete:
            operationId: deleteObserverRequest
            parameters:
                - in: path
                  name: observer
                  required: true
                  type: string
                  x-go-name: Observer
            responses:
                "200":
                    $ref: '#/responses/deleteObserverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete observer by name.
            tags:
                - Observer
        put:
            operationId: updateObserverRequest
            parameters:
                - in: path
                  name: observer
                  required: true
                  type: string
                  x-go-name: Observer
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ObserverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateObserverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update observer by name, the observer must already exist.
            tags:
                - Observer
    /config/recorders:
        post:
            operationId: createRecorderRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RecorderConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createRecorderResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new recorder, the name of the recorder must be unique in recorder list.
            tags:
                - Recorder
    /config/recorders/{recorder}:
        delete:
            operationId: deleteRecorderRequest
            parameters:
                - in: path
                  name: recorder
                  required: true
                  type: string
                  x-go-name: Recorder
            responses:
                "200":
                    $ref: '#/responses/deleteRecorderResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete recorder by name.
            tags:
                - Recorder
        put:
            operationId: updateRecorderRequest
            parameters:
                - in: path
                  name: recorder
                  required: true
                  type: string
                  x-go-name: Recorder
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RecorderConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateRecorderResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update recorder by name, the recorder must already exist.
            tags:
                - Recorder
...
    /config/rlimiters:
        post:
            operationId: createRateLimiterRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new rate limiter, the name of limiter must be unique in limiter list.
            tags:
                - Limiter
    /config/rlimiters/{limiter}:
        delete:
            operationId: deleteRateLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
            responses:
                "200":
                    $ref: '#/responses/deleteRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete rate limiter by name.
            tags:
                - Limiter
        put:
            operationId: updateRateLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateRateLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update rate limiter by name, the limiter must already exist.
            tags:
                - Limiter
...
produces:
    - application/json
responses:
...
    createObserverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createRecorderResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteObserverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteRecorderResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateObserverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateRateLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateRecorderResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"